bug修改，发货修改,发货数量统计，批购发货弹窗样式调整、批购详情跨行合并样式（未完）。4h

+++ src/main/java/com/comdosoft/financial/manage/controller/order/OrderAgentController.java
@@ -156,7 +156,7 @@ public class OrderAgentController extends BaseController {
 
 	@RequestMapping(value = "/agent/{id}/save", method = RequestMethod.GET)
 	public String save(HttpServletRequest request,@PathVariable Integer id, Byte status,
-			Integer actualPrice, Model model) {
+			Float actualPrice, Model model) {
 		orderService.save(id, status, actualPrice, null);
 		Order order = orderService.findOrderInfo(id);
 		model.addAttribute("order", order);
@@ -172,7 +172,7 @@ public class OrderAgentController extends BaseController {
 	
 	@RequestMapping(value = "/agent/info/{id}/save", method = RequestMethod.GET)
 	public String saveInfo(HttpServletRequest request,@PathVariable Integer id, Byte status,
-			Integer actualPrice, Model model) {
+			Float actualPrice, Model model) {
 		orderService.save(id, status, actualPrice, null);
 		Order order = orderService.findOrderInfo(id);
 		model.addAttribute("order", order);
